requirements.txt
asgiref==3.2.7
Django==3.0.4
django-markdown==0.8.4
django-markdownx==3.0.1
Markdown==3.2.1
Pillow==7.1.1
Pygments==2.6.1
pytz==2019.3
sqlparse==0.3.1
